WORKING SPAWN WALL SYSTEM
- added more comments, removing Debug.Log from mirror source

- Really good clean up on GunScript where a lot of that stuff (gold, ammo, health) is now in the InGameScript - this was something I needed to do for a while, code is much cleaner as well

- all input is now InGameScript Update() method, potential bug as players can't hold down the shooting key and spray bullets, they have to keep clicking to shoot one bullet at a time

- implemented "Defence Spawner" system, where an instance of it is created from NetworkManager and has DontDestoryOnLoad() so it loads into game. When players want to spawn a wall/foundation, the prefab index (index from defence spawner) is sent to the server along with its Vector3 position and Quaterion rotation. This information is used on the server where the server grabs the index, uses index from DefenceSpawner created in the game, and instantiates it using the position and rotation. NetworkServer.Spawn(..) is then used to spawn the object for all players. Previews are kept local and are destroyed locally.
