+++ Resistance/Assets/Resources/Spawnable/Prefabs/old/INVENTORY.prefab
@@ -285,7 +285,7 @@ GameObject:
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
   m_StaticEditorFlags: 0
-  m_IsActive: 1
+  m_IsActive: 0
 --- !u!4 &4968830939164728261
 Transform:
   m_ObjectHideFlags: 0
@@ -314,6 +314,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: cd821fefe02eeff47a8090976bc62d5b, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  syncMode: 0
+  syncInterval: 0.1
   layer:
     serializedVersion: 2
     m_Bits: 3383
@@ -400,8 +402,8 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   foundationPreview: {fileID: 0}
   wallPreview: {fileID: 0}
+  inventoryUI: {fileID: 0}
   buildSystem: {fileID: 0}
-  playerInventory: {fileID: 0}
 --- !u!114 &1856255112906718208
 MonoBehaviour:
   m_ObjectHideFlags: 0
