+++ Resistance/Assets/Resources/Spawnable/Structure/Foundation/Foundation preview.prefab
@@ -14,6 +14,7 @@ GameObject:
   - component: {fileID: -6937602260139318547}
   - component: {fileID: -264650420278728531}
   - component: {fileID: -4118077946869901380}
+  - component: {fileID: 2842093065700627566}
   m_Layer: 9
   m_Name: Foundation preview
   m_TagString: primitive
@@ -130,3 +131,20 @@ MonoBehaviour:
   isFoundation: 1
   tagsToSnapTo:
   - SP_Foundation
+--- !u!114 &2842093065700627566
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 411865924816708581}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9b91ecbcc199f4492b9a91e820070131, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  sceneId: 0
+  serverOnly: 0
+  m_AssetId: 
+  _allowChildrenBehaviours: 0
+  hasSpawned: 0
