+++ Resistance/Assets/Resources/Spawnable/Structure/Walls/Castle Wall 1 Variant.prefab
@@ -1,5 +1,22 @@
 %YAML 1.1
 %TAG !u! tag:unity3d.com,2011:
+--- !u!114 &-78587115346591843
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8357639209293913169}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9b91ecbcc199f4492b9a91e820070131, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  sceneId: 0
+  serverOnly: 0
+  m_AssetId: 
+  _allowChildrenBehaviours: 0
+  hasSpawned: 0
 --- !u!1001 &1593530930412307018
 PrefabInstance:
   m_ObjectHideFlags: 0
@@ -436,6 +453,12 @@ PrefabInstance:
       objectReference: {fileID: 0}
     m_RemovedComponents: []
   m_SourcePrefab: {fileID: 100100000, guid: 4f1378d27988e7148ae48e6635728e82, type: 3}
+--- !u!1 &8357639209293913169 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 1106412074651284, guid: 4f1378d27988e7148ae48e6635728e82,
+    type: 3}
+  m_PrefabInstance: {fileID: 8358578456861627077}
+  m_PrefabAsset: {fileID: 0}
 --- !u!4 &8353838649225957209 stripped
 Transform:
   m_CorrespondingSourceObject: {fileID: 4900727511227804, guid: 4f1378d27988e7148ae48e6635728e82,
