+++ Resistance/Assets/Resources/Spawnable/Structure/Walls/Castle Wall 1 preview.prefab
@@ -29,11 +29,29 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   prefab: {fileID: 8357639209293913169, guid: abbcd0b3a5c231b47bdd1f72ed036010, type: 3}
+  buildSystem: {fileID: 0}
   legalMat: {fileID: 0}
   illegalMat: {fileID: 0}
   isFoundation: 0
   tagsToSnapTo:
   - SP_WALL
+--- !u!114 &-8694911596012403809
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 6978302913536196828}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9b91ecbcc199f4492b9a91e820070131, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  sceneId: 0
+  serverOnly: 0
+  m_AssetId: 
+  _allowChildrenBehaviours: 0
+  hasSpawned: 0
 --- !u!1001 &6977210245441998408
 PrefabInstance:
   m_ObjectHideFlags: 0
