+++ Resistance/Assets/Resources/Spawnable/Structure/Walls/Castle Wall 3 preview Variant.prefab
@@ -32,6 +32,23 @@ MonoBehaviour:
   isFoundation: 0
   tagsToSnapTo:
   - SP_WALL
+--- !u!114 &6301305176857390455
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1617765394855863686}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9b91ecbcc199f4492b9a91e820070131, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  sceneId: 0
+  serverOnly: 0
+  m_AssetId: 
+  _allowChildrenBehaviours: 0
+  hasSpawned: 0
 --- !u!1001 &1618717033044177750
 PrefabInstance:
   m_ObjectHideFlags: 0
