+++ Resistance/Assets/Resources/Spawnable/Structure/Walls/Castle Wall 8 Variant.prefab
@@ -13,6 +13,23 @@ BoxCollider:
   serializedVersion: 2
   m_Size: {x: 6.75, y: 3.7500005, z: 0.75001}
   m_Center: {x: 0, y: 1.8749993, z: -0.000005155802}
+--- !u!114 &6794405345217198701
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 8517660160276756776}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9b91ecbcc199f4492b9a91e820070131, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  sceneId: 0
+  serverOnly: 0
+  m_AssetId: 
+  _allowChildrenBehaviours: 0
+  hasSpawned: 0
 --- !u!1001 &599704441310266160
 PrefabInstance:
   m_ObjectHideFlags: 0
