+++ Resistance/Assets/Resources/Spawnable/Structure/Walls/Castle Wall 8 preview Variant.prefab
@@ -32,6 +32,23 @@ MonoBehaviour:
   isFoundation: 0
   tagsToSnapTo:
   - SP_WALL
+--- !u!114 &-8133834307549890205
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2690805022914838452}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 9b91ecbcc199f4492b9a91e820070131, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  sceneId: 0
+  serverOnly: 0
+  m_AssetId: 
+  _allowChildrenBehaviours: 0
+  hasSpawned: 0
 --- !u!1001 &2689693206602534190
 PrefabInstance:
   m_ObjectHideFlags: 0
