+++ Resistance/Assets/Resources/SpawnablePrefabs/DefenceSpawner.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: 9d8bc9729cfff8f45b6769b3a800ecbf
-folderAsset: yes
-DefaultImporter:
+guid: 92ee6ea737d652d44a31ad3ce398484e
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
