+++ Resistance/Assets/Resources/SpawnablePrefabs/Foundation.prefab.meta
@@ -1,7 +1,6 @@
 fileFormatVersion: 2
-guid: d9fcae517fd99b340a61e1790c571962
-folderAsset: yes
-DefaultImporter:
+guid: 4e768d9cf3053e3458efffc5b6609cb0
+PrefabImporter:
   externalObjects: {}
   userData: 
   assetBundleName:
