+++ Resistance/Assets/Scenes/MainScene.unity
@@ -38,7 +38,7 @@ RenderSettings:
   m_ReflectionIntensity: 1
   m_CustomReflection: {fileID: 0}
   m_Sun: {fileID: 0}
-  m_IndirectSpecularColor: {r: 0.0020952, g: 0.00081933395, b: 0.002809898, a: 1}
+  m_IndirectSpecularColor: {r: 0.0021282036, g: 0.0008354882, b: 0.0028883044, a: 1}
   m_UseRadianceAmbientProbe: 0
 --- !u!157 &3
 LightmapSettings:
@@ -853,7 +853,7 @@ PrefabInstance:
     - target: {fileID: 2723663690555767445, guid: dd691567504cccf40b5b78e0f990565a,
         type: 3}
       propertyPath: m_RootOrder
-      value: 2
+      value: 3
       objectReference: {fileID: 0}
     - target: {fileID: 2723663690555767445, guid: dd691567504cccf40b5b78e0f990565a,
         type: 3}
@@ -912,7 +912,7 @@ Transform:
   - {fileID: 592679551}
   - {fileID: 1555504150}
   m_Father: {fileID: 0}
-  m_RootOrder: 0
+  m_RootOrder: 1
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1555504149
 GameObject:
@@ -1333,7 +1333,7 @@ PrefabInstance:
     - target: {fileID: 3563727116531817320, guid: 17e113675001f08499393b8ebfde0d0d,
         type: 3}
       propertyPath: m_RootOrder
-      value: 1
+      value: 2
       objectReference: {fileID: 0}
     - target: {fileID: 3563727116531817320, guid: 17e113675001f08499393b8ebfde0d0d,
         type: 3}
@@ -1756,87 +1756,3 @@ RectTransform:
     type: 3}
   m_PrefabInstance: {fileID: 5026931744928672886}
   m_PrefabAsset: {fileID: 0}
---- !u!1001 &7161378142115513816
-PrefabInstance:
-  m_ObjectHideFlags: 0
-  serializedVersion: 2
-  m_Modification:
-    m_TransformParent: {fileID: 0}
-    m_Modifications:
-    - target: {fileID: 1285652833997433759, guid: 21090972173c36541b98f364b68bea54,
-        type: 3}
-      propertyPath: m_AnchorMax.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 1285652833997433759, guid: 21090972173c36541b98f364b68bea54,
-        type: 3}
-      propertyPath: m_AnchorMax.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 5546782658093222849, guid: 21090972173c36541b98f364b68bea54,
-        type: 3}
-      propertyPath: m_LocalPosition.x
-      value: 178.08
-      objectReference: {fileID: 0}
-    - target: {fileID: 5546782658093222849, guid: 21090972173c36541b98f364b68bea54,
-        type: 3}
-      propertyPath: m_LocalPosition.y
-      value: 0.5
-      objectReference: {fileID: 0}
-    - target: {fileID: 5546782658093222849, guid: 21090972173c36541b98f364b68bea54,
-        type: 3}
-      propertyPath: m_LocalPosition.z
-      value: -21.64
-      objectReference: {fileID: 0}
-    - target: {fileID: 5546782658093222849, guid: 21090972173c36541b98f364b68bea54,
-        type: 3}
-      propertyPath: m_LocalRotation.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 5546782658093222849, guid: 21090972173c36541b98f364b68bea54,
-        type: 3}
-      propertyPath: m_LocalRotation.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 5546782658093222849, guid: 21090972173c36541b98f364b68bea54,
-        type: 3}
-      propertyPath: m_LocalRotation.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 5546782658093222849, guid: 21090972173c36541b98f364b68bea54,
-        type: 3}
-      propertyPath: m_LocalRotation.w
-      value: 1
-      objectReference: {fileID: 0}
-    - target: {fileID: 5546782658093222849, guid: 21090972173c36541b98f364b68bea54,
-        type: 3}
-      propertyPath: m_RootOrder
-      value: 9
-      objectReference: {fileID: 0}
-    - target: {fileID: 5546782658093222849, guid: 21090972173c36541b98f364b68bea54,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.x
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 5546782658093222849, guid: 21090972173c36541b98f364b68bea54,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.y
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 5546782658093222849, guid: 21090972173c36541b98f364b68bea54,
-        type: 3}
-      propertyPath: m_LocalEulerAnglesHint.z
-      value: 0
-      objectReference: {fileID: 0}
-    - target: {fileID: 8836476609283717586, guid: 21090972173c36541b98f364b68bea54,
-        type: 3}
-      propertyPath: m_Name
-      value: testingprefab
-      objectReference: {fileID: 0}
-    - target: {fileID: 8836476609283717586, guid: 21090972173c36541b98f364b68bea54,
-        type: 3}
-      propertyPath: m_TagString
-      value: Player
-      objectReference: {fileID: 0}
-    m_RemovedComponents: []
-  m_SourcePrefab: {fileID: 100100000, guid: 21090972173c36541b98f364b68bea54, type: 3}
