+++ Resistance/Assets/Scripts/Input/InGameScript.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 872f8d75813de3243846b75da0e3025e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
