+++ Resistance/Assets/Scripts/Lobby Scripts/NetworkManagerLobby.cs
@@ -17,6 +17,7 @@ public class NetworkManagerLobby : NetworkManager
     [Header("Game")]
     [SerializeField] private NetworkGamePlayerLobby gamePlayerPrefab = null;
     [SerializeField] private GameObject playerSpawnSystem = null;
+    [SerializeField] private DefenceSpawner defenceSpawner = null;
     // [SerializeField] private GameObject roundSystem = null;
 
     private int playerIndexes = 0;
@@ -165,6 +166,7 @@ public class NetworkManagerLobby : NetworkManager
                 NetworkServer.Destroy(conn.identity.gameObject);
                 NetworkServer.ReplacePlayerForConnection(conn, gameplayerInstance.gameObject, true);
             }
+            var ds = Instantiate(defenceSpawner);
         }
         base.ServerChangeScene(newSceneName);
     }
