+++ Resistance/Assets/Tests/EditMode/EditMode.cs
@@ -9,18 +9,18 @@ namespace Tests
 {
     public class EditMode
     {
-        [Test]
-        public void UI_Information() //Unit Test for UI info
-        {
-            var gameObject = new GameObject();
-            var UI_Info = gameObject.AddComponent<GunScript>();
+        //[Test]
+        //public void UI_Information() //Unit Test for UI info
+        //{
+        //    var gameObject = new GameObject();
+        //    var UI_Info = gameObject.AddComponent<GunScript>();
 
-            Assert.AreEqual(20, UI_Info.maxClipAmmo);
-            Assert.AreEqual(100, UI_Info.maxTotalAmmo);
+        //    Assert.AreEqual(20, UI_Info.maxClipAmmo);
+        //    Assert.AreEqual(100, UI_Info.maxTotalAmmo);
 
-            Assert.AreEqual(1000, UI_Info.startingGold);
+        //    Assert.AreEqual(1000, UI_Info.startingGold);
 
-            Assert.AreEqual(100, UI_Info.maxHealth);
-        }
+        //    Assert.AreEqual(100, UI_Info.maxHealth);
+        //}
     }
 }
